+++ src/gamelogic/rapidXML/RapidXMLAdapter.cpp
@@ -149,7 +149,7 @@ namespace GameLogic {
     }
 
 
-    string RapidXMLAdapter::get_entity_texture( int x, int y )
+    string RapidXMLAdapter::get_entity_texture( size_t x, size_t y )
     {
         for ( xml_node<> * object_node = this->_object_group_node->first_node( "object" ); object_node; object_node = object_node->next_sibling() ) {
             if ( std::stoi( object_node->first_attribute( "x" )->value() ) / this->_tile_width == x &&  std::stoi( object_node->first_attribute( "y" )->value() ) / this->_tile_height == y) {
